Provide the correct `rootpath` and `uri` for special pages.
When rendering special pages like the masterindex, no document content
is available. Instead, a synthetic rendering model is created. This
model did not provide access to the `uri` of the rendered page and had
not the correct `rootpath` set (at least if the special pages are mapped
to a non-root path).